NotifierTest: Refactor, relocating ModelConditionRecorder support to base class
